Tweaked "Always on top option"
"Always on top" used to make all windows, including the
file dialog of the setup window's "Add..." button modal.
The file dialog was then not able to accept dropped files.

If the setup window is shown, the main window becomes "normal"
and only when the setup window is hidden again, it becomes
either "modal" or remains "normal", depending on the "Always
on top setting".

Fixed the one-click-to-launch message.

Updated docs.