Fixed MouseUp() for single-click
Mouse up doesn't reliable report which mouse button was releases.
So we remember if it was the primary button in MouseDown() with
fPrimaryButton and also which item was clicked with fCurrentItemIndex.

In MouseUp() we check if it was the primary button and still the same
item. Only then send the SINGLE_CLICK message.

A few simplicications.

Thanks AnEvilYak for his keen eye, once more...