Don't save favorites as fragile entry_refs
Turns out, entry_refs are very fragile and therefore a bad choice for
saving a list of favorites.

The device ID can change when booting from a USB device, the package-fs
being virtual will also keep changin the inode numbers, etc. Updating
packages will therefore change entries and we won't find our favourite
apps anymore.

Using absolute paths as strings now.
The alternative would be saving app signatures, but that leads to problems
when more versions of an app are present in the system. Not uncommon for
developers IMO.
The downside of paths is that if a package changes the app's name or
install path, it'll break and the user has to re-add the favorite...
I consider this not to be the case often.