Favorites now drag&dropable, too
Use B_SIMPLE_DATA for dragged favorites as well and add a "isfav"
bool to the message for identification.
Now favorites can be dragged to e.g. LaunchBox just like apps from
QuickLaunch's result list.

Positive side effects:
The position of a favorite can't be changed via drag&drop in a result
list, only in the favorites list.
Favorites can be added to the ignore list of the Setup window. They then
only appear in the favorites list, not the result list.

Fixes #13.