Select the app after opening its containing folder
Lifting code from opentargetfolder :)
Caveat: if the app is in a very crowded folder like /bin, Tracker
may take too long populating the window. Then the message to select
the item arrives before the item was added to the window and fails.

To improve this situation slightly, we snooze for 0.3 secs to give
Tracker time to populate the Window.

Fixes #14.