Solve drag'n'drop conundrum with modal windows
Being modal windows, you're unable to browse the file
system to drag'n'drop files into the ignore list. Worse,
the file dialog that opens when clicking the Add... button
isn't reachable either.

The work around now is that while the Setup window is open,
the 'modal' main window becomes 'normal'.

Thanks Diver for reporting this bug. Apparently I haven't used
the file dialog or tried to d'n'd files in a long time. I usually
add to-be-ignored files via the context menu of the main window...