Temporary: Check if there's a Shortcuts prefs with old QL location
Since we move the QuickLaunch binary from /system/apps/QuickLaunch/
to /system/apps/, we check if there's an entry for QuickLaunch in the
Shortcuts preferences file.
If so we put up an alert, offering to open the Shortcuts prefs so the user
can enter the correct location.

Can be reverted once we can assume most users have migrated after
Haiku R1beta5.