Support localized application names
A query is only able to to find the English original name of apps.

Instead of doing a query every time the user enters a letter, we now do one
query for all applications (minus the ones from the ignore list) when creating
the MainWindow (see BuildAppList()) and put them into fAppList.

If "Translate app names" is activated in Haiku's Locale prefs, we use the
localized filenames instead of the regular names.

When a letter is entered, we empty the MainListView and only add the matching
items from fAppList (see FilterAppList(), previously named BuildList()).

When the ignore list changes or is de/activated we empty fAppList and call
BuildAppList() to create a new one.