Make window "float" instead of "modal"
The main window used to be modal, i.e. always on top and having the
focus. Now it just floats on top, i.e. it can lose focus and the user
can interact with other apps.

This is necessary, because input methods like Mozc for Japanese try
to show a window with suggestions that is obscured by a modal QuickLaunch
window.

Bonus: Now an app item can be drag&dropped, for example into LaunchBox.

This whould be the last change to close issue #41.

Fixes #41
Fixes #24
Fixes #13