BuildAppList in its own thread
If there are many applications, maybe from several Haiku installs
on a number of mounted partitions, possible on slower media, building
the list of all apps takes time.
This is especially annoying if the window isn't shown until the list
has finished building. Feels sluggish.

We now build the app list in its own thread. This allows showing the
window at once, with the favorites that can be invoked without having
to wait on the app list.