Don't send NEW_FILTER message on key press
Since we get a ModificationMessage when the search control contents
changes, the key-catching filter doesn't have to forward keypresses
anymore.

By doing that, we actually filtered the list twice with every entered
letter!
This also resulted in a graphical glitch when there are many
apps added to the list, like when entering "*" which shows all apps:
The window was resized at the end of the first filtering, and immediately
the list was emptied for the second round of filtering, leaving the
empty resized part for a fraction of a second while the list was being
filtered...